Type changes

+++ app/src/main/cpp/MediTrakCore/Dose/Dose.cpp
@@ -12,7 +12,15 @@ Dose::Dose() {
     doseTime = {};
 }
 
-Dose::Dose(long id, long medicationId, bool taken, std::string doseTime, std::string timeTaken, int overrideDoseAmount, std::string overrideDoseUnit) {
+Dose::Dose(
+        long id,
+        long medicationId,
+        bool taken,
+        std::string doseTime,
+        std::string timeTaken,
+        float overrideDoseAmount,
+        std::string overrideDoseUnit
+) {
     this->id = id;
     this->medicationId = medicationId;
     this->taken = taken;
